Finally got around to uploading old projects
These files are exactly what was turned in to professor for a grade, so things are pretty minimal. The textfile README was supposed to only list what parts of the project had been completed, what was an issue, etc. The report.txt file was also minimal, only explaining the basics of our desinged protocol and the structure of our packets.

Actual documentation and a guide for the project (written by myself) is found in the README.md file